Closes trac:1118: 32 bit shared memory size limit fix
Submitted by paklui, Reviewed by Gleb, RM-Approved by Terry

r15472: Remove a compilation warning.

r15752:
 * restrict the user with 32 bit app to specify a sm_size to be between
   2GB to 4GB-1 by using long instead of size_t for the sm size.
 * it is done to prevent user from running into the ftruncate() in
   common sm component (and possibly others) problem that ftruncate
   takes an off_t which is a signed long integer. If we use an
   unsigned long, it'll run into an invalid argument errno=22.
 * See trac #1117

r15790:
Update the comment for the previous fix. Thanks Gleb for pointing out.

This commit was SVN r15810.

The following SVN revision numbers were found above:
  r15472 --> open-mpi/ompi@59ee366728c348381e38b48078e47e1a313bfba3
  r15752 --> open-mpi/ompi@010d216db90576b84b7d1a957a7680986b506942
  r15790 --> open-mpi/ompi@0790c4cc405a65510e33dd4e979caee2f563983b

The following Trac tickets were found above:
  Ticket 1118 --> https://svn.open-mpi.org/trac/ompi/ticket/1118

+++ ompi/mca/mpool/sm/mpool_sm.h
@@ -36,7 +36,7 @@ struct mca_mpool_sm_component_t {
   mca_mpool_base_component_t super;
   /*      mca_allocator_base_module_t* sm_allocator; */
   char*  sm_allocator_name;
-  size_t sm_size;
+  long   sm_size;
   int    verbose;
   /*      struct mca_mpool_sm_mmap_t *sm_mmap; */
 };
